Adds new RequestDataCollectorInterface and moves some methods to BaseGUIRequest
